made it into a Stack package, formatted, and added documentation
